Renaming the request_source_plugin to request_source_plugin_config_factory. Signed-off-by: William Juan <66322422+wjuan-AFK@users.noreply.github.com>
