+++ source/common/request_source_plugin_impl.h
@@ -2,7 +2,7 @@
 
 #include "envoy/registry/registry.h"
 
-#include "nighthawk/common/request_source_plugin.h"
+#include "nighthawk/common/request_source_plugin_config_factory.h"
 
 #include "external/envoy/source/common/common/lock_guard.h"
 #include "external/envoy/source/common/common/thread.h"
